introduced additional benchmarking tests
